% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{engine_capabilities}
\alias{engine_capabilities}
\title{Inspect colocalization-engine capabilities}
\usage{
engine_capabilities(engine, ...)
}
\arguments{
\item{engine}{A colocalization engine.}

\item{...}{Engine-specific arguments.}
}
\value{
A named list.
}
\description{
Inspect colocalization-engine capabilities
}
