% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{estimate_weights}
\alias{estimate_weights}
\title{Estimate ancestry weights}
\usage{
estimate_weights(provider, sumstats, locus = NULL, dataset_id = NULL, ...)
}
\arguments{
\item{provider}{Object implementing the \code{AncestryWeightsProvider} interface.}

\item{sumstats}{A \link{HarmonizedSumstats} object or compatible table.}

\item{locus}{Optional \link{Locus}.}

\item{dataset_id}{Optional dataset ID.}

\item{...}{Backend-specific arguments.}
}
\value{
An \link{AncestryWeights} object.
}
\description{
Estimate ancestry weights
}
