% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{finemap_schema}
\alias{finemap_schema}
\title{Inspect a fine-mapping provider schema}
\usage{
finemap_schema(provider, ...)
}
\arguments{
\item{provider}{Object implementing the \code{FineMapResultProvider} interface.}

\item{...}{Backend-specific arguments.}
}
\value{
Provider-specific schema metadata.
}
\description{
Inspect a fine-mapping provider schema
}
