% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{FineMapResultProvider}
\alias{FineMapResultProvider}
\title{Fine-mapping result provider contract}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
FineMapResultProvider
}
\description{
\code{FineMapResultProvider} is for tools whose coloc layer consumes fine-mapping
outputs rather than raw summary statistics, such as SuSiEx/MsCAVIAR outputs
used by MACA.
}
\keyword{datasets}
