% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{normalize_result}
\alias{normalize_result}
\title{Normalize a raw colocalization result}
\usage{
normalize_result(engine, raw_result, ...)
}
\arguments{
\item{engine}{A colocalization engine.}

\item{raw_result}{Raw engine output.}

\item{...}{Engine-specific arguments.}
}
\value{
A \link{ColocResult}.
}
\description{
Normalize a raw colocalization result
}
