% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostGWAS-package.R
\docType{package}
\name{PostGWAS-package}
\alias{PostGWAS}
\alias{PostGWAS-package}
\title{PostGWAS: African-ancestry-first post-GWAS workflows}
\description{
PostGWAS depends on GAUSS.pgen for native PLINK2 \code{.pgen} LD,
summary-statistic imputation, \code{.psam} population keep-file utilities, and
unrelated-sample keep-file derivation. Higher-level post-GWAS workflows should
build on those GAUSS.pgen primitives instead of duplicating them here.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/PostGWAS}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/PostGWAS/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Other contributors:
\itemize{
  \item Donghyung Lee \email{leed13@miamioh.edu} (Original GAUSS summary-statistic methods) [contributor]
}

}
\keyword{internal}
