% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{query_finemap_results}
\alias{query_finemap_results}
\title{Query fine-mapping results}
\usage{
query_finemap_results(provider, locus, traits = NULL, features = NULL, ...)
}
\arguments{
\item{provider}{Object implementing the \code{FineMapResultProvider} interface.}

\item{locus}{A \link{Locus}.}

\item{traits}{Optional trait IDs.}

\item{features}{Optional feature IDs.}

\item{...}{Backend-specific arguments.}
}
\value{
A \link{FineMapResult}.
}
\description{
Query fine-mapping results
}
