% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{query_sumstats}
\alias{query_sumstats}
\title{Query summary statistics}
\usage{
query_sumstats(
  provider,
  locus,
  traits = NULL,
  features = NULL,
  columns = NULL,
  ...
)
}
\arguments{
\item{provider}{Object implementing the \code{SumstatProvider} interface.}

\item{locus}{A \link{Locus}.}

\item{traits}{Optional trait IDs to retain.}

\item{features}{Optional feature IDs to retain.}

\item{columns}{Optional columns to project.}

\item{...}{Backend-specific arguments.}
}
\value{
A \link{HarmonizedSumstats} object.
}
\description{
Query summary statistics
}
