% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{resource_metadata}
\alias{resource_metadata}
\title{Inspect provider resource metadata}
\usage{
resource_metadata(provider, ...)
}
\arguments{
\item{provider}{Any resource provider.}

\item{...}{Backend-specific arguments.}
}
\value{
A list of metadata.
}
\description{
Inspect provider resource metadata
}
