% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run_coloc}
\alias{run_coloc}
\title{Run a colocalization engine}
\usage{
run_coloc(engine, bundle, ...)
}
\arguments{
\item{engine}{Object implementing the \code{ColocEngine} interface.}

\item{bundle}{A \link{ColocInputBundle}.}

\item{...}{Engine-specific arguments.}
}
\value{
A \link{ColocResult}.
}
\description{
Run a colocalization engine
}
