% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{SignedCorrelationLD}
\alias{SignedCorrelationLD}
\title{Signed correlation LD trait}
\format{
An object of class \code{s7contract::s7_trait} (inherits from \code{S7_object}) of length 1.
}
\usage{
SignedCorrelationLD
}
\description{
Explicit capability marker for LD objects/providers that promise signed LD on
the correlation scale. This semantic promise is stronger than merely having a
method named \code{compute_ld()}.
}
\keyword{datasets}
