% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interfaces.R
\docType{data}
\name{SumstatProvider}
\alias{SumstatProvider}
\title{Summary-statistic provider contract}
\format{
An object of class \code{s7contract::s7_interface} (inherits from \code{S7_object}) of length 1.
}
\usage{
SumstatProvider
}
\description{
\code{SumstatProvider} is a small behavioral contract for objects that can return
harmonized summary statistics for a locus. Concrete implementations may read
data frames, files, DuckDB relations, Rduckhts/Rducks tables, or remote stores.
}
\keyword{datasets}
